// frontend/src/app/types/api.d.ts
export type components = {
    schemas: {
// ...
        ReservationDataResponse: components["schemas"]["BaseResponse"] & {
            data?: components["schemas"]["Reservation"];
        };
        ReservationGetInput: {
            fields?: ("id" | "clientId" | "roomId" | "status" | "checkInDate" | "checkOutDate" | "charge" | "active" | "creationDate")[];
            reservationIds: number[];
        };
        ReservationGetResponse: components["schemas"]["BaseResponse"] & {
            data?: components["schemas"]["Reservation"][];
        };
// ...
        ReservationListResponse: components["schemas"]["BaseResponse"] & {
            data?: components["schemas"]["Reservation"][];
            page?: number;
            total?: number;
        };
// ...
        ResultResponse: components["schemas"]["BaseResponse"] & {
            result?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: {
// ...
        ReservationCreateRequest: {
            content: {
                "application/json": components["schemas"]["ReservationCreateInput"];
            };
        };
        ReservationDeleteRequest: {
            content: {
                "application/json": {
                    reservationId: number;
                };
            };
        };
        ReservationGetRequest: {
            content: {
                "application/json": components["schemas"]["ReservationGetInput"];
            };
        };
        ReservationListRequest: {
            content: {
                "application/json": components["schemas"]["ReservationListInput"];
            };
        };
        ReservationRestoreRequest: {
            content: {
                "application/json": {
                    reservationId: number;
                };
            };
        };
        ReservationUpdateRequest: {
            content: {
                "application/json": components["schemas"]["ReservationUpdateInput"];
            };
        };
// ...
    };
    headers: never;
    pathItems: never;
};